feat(raft): paginate the unapplied config changes scan (#2)
Fix potentially unlimited memory usage spike possible in raft.hup()
which reads all unapplied committed entries in order to check that there
are no unapplied config changes. This PR paginates this scan so that the
spike is limited to MaxCommittedSizePerReady. It also terminates the
scan early if a config change has been found.

It is ported from etcd-io/raft#32.

It is ported from tikv/raft-rs#530.

Signed-off-by: tison <wander4096@gmail.com>